refactor(🚇 FilterRegistry): 🎣 HookRegistry is now 🚇 FilterRegistry
