+++ COMMIT.md
@@ -4,7 +4,7 @@ Emoji reference of Commitizen scopes for this repo:
 📔 Logger
 📦 BlueBaseModule
 🎁 ComponentRegistry
-🎣 HookRegistry
+🚇 FilterRegistry
 🗿 Icons
 🍱 JSON Schema
 🔌 PluginRegistry
@@ -16,7 +16,7 @@ Emoji reference of Commitizen scopes for this repo:
 
 - 🎁 ComponentRegistry
 - 🎛 ConfigRegistry
-- 🎣 HookRegistry
+- 🚇 FilterRegistry
 - 🔌 PluginRegistry
 - 🎨 ThemeRegistry
