+++ src/registries/index.ts
@@ -1,6 +1,6 @@
 export * from './ComponentRegistry';
 export * from './ConfigRegistry';
-export * from './HookRegistry';
+export * from './FilterRegistry';
 export * from './PluginRegistry';
 export * from './ThemeRegistry';
 export * from './Registry';
